pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Heloisa", email = "heloisa@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.0.0",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
